springlens-ragas-evaluator/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
ragas==0.2.15
langchain-openai==0.3.9
pydantic==2.10.4
